requirements.txt
python-dotenv
requests
airtable-python-wrapper
google-generativeai
